refactor: make code better

+++ src/resources/users/user.router.ts
@@ -2,10 +2,8 @@ import express, {Router, Request, Response, NextFunction} from 'express';
 import * as bodyParser from 'body-parser';
 import User from './user.model';
 import * as usersService from './user.service';
-import getStatus from '../../utils/router.helpers';
-
 import { IUser } from '../../types';
-// import { NextFunction } from 'express';
+
 
 
 const router: Router = express.Router();
@@ -13,9 +11,10 @@ router.use(bodyParser.text());
 
 router.route('/').get(async (req:Request, res:Response, next:NextFunction) => {
   try {
-    const users = await usersService.getAll()
+    const users = await usersService.getAll();
+    if (!users) throw new Error("NOO users")
     res.json(users.map(User.toResponse));
-    // throw new Error("MY OWN ERROR")
+    // 
   } catch(err) {
     next(err)
   }
@@ -27,6 +26,7 @@ router
     try{ 
       const { id } = req.params;
     const user = await usersService.getUser(id);
+    if (!user || !id) throw new Error("NOO users or id");
     res.status(200).json(user);
     }catch(err){
       next(err)
@@ -36,6 +36,7 @@ router
   .post('/', async (req:Request, res: Response, next:NextFunction) => {
 try{
     const user: IUser | string = await usersService.addUser(req.body);
+    if (!user) throw new Error("NOO users ");
     res.status(201).json(user);
   }catch(err){
     next(err)
@@ -46,6 +47,7 @@ try{
     try{
     const { id } = req.params;
     const user: IUser | string  = await usersService.updateUser(id, req.body);
+    if (!user || !id) throw new Error("NOO users or id");
     res.status(200).json(user);
   }catch(err){
     next(err)
@@ -55,6 +57,7 @@ try{
   .delete('/:id', async (req:Request, res:Response, next:NextFunction) => {
     try{
     const { id } = req.params;
+    if ( !id) throw new Error("NOO id");
     const result:number | string = await usersService.deleteUser(id);
     res.status(200).json(result);
   }catch(err){
